renamed LoginActivity to SignupActivity, created another LoginActivity to act as a true LOGIN page
