+++ app/src/main/AndroidManifest.xml
@@ -6,24 +6,22 @@
     <uses-permission android:name="android.permission.INTERNET" />
 
     <application
-
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.ThoughtWell">
-
-        <!-- Login activity will be launched on app first startup, this config below assures that.-->
-        <activity android:name=".LoginActivity">
+        <activity android:name=".LoginActivity"></activity>
+        <!-- Login activity will be launched on app first startup, this config below assures that. -->
+        <activity android:name=".SignupActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
+
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-
-        <activity android:name=".MainActivity"></activity>
-
+        <activity android:name=".MainActivity" />
     </application>
 
 </manifest>
