+++ app/src/main/java/com/example/thoughtwell/MainActivity.java
@@ -5,11 +5,11 @@ import android.os.Bundle;
 import android.util.Log;
 import android.view.View;
 import android.widget.Button;
-import android.widget.EditText;
 import android.widget.TextView;
 
 import androidx.appcompat.app.AppCompatActivity;
 
+import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
 
@@ -20,7 +20,8 @@ public class MainActivity extends AppCompatActivity {
 
     //private Button btnNext;
     //private Button btnDeposit;
-    private Button btnLogin;
+    //private Button btnLogin;
+    private Button btnLogout;
     private TextView tvSampleThought;
 
     FirebaseDatabase database = FirebaseDatabase.getInstance();
@@ -37,15 +38,39 @@ public class MainActivity extends AppCompatActivity {
         // get references to buttons
         //btnNext = findViewById(R.id.btnNext);
         // btnDeposit = findViewById(R.id.btnDeposit);
-        btnLogin = findViewById(R.id.btnLogin);
+        //btnLogin = findViewById(R.id.btnLogin);
+        btnLogout = findViewById(R.id.btnLogout);
         tvSampleThought = findViewById(R.id.tvSampleThought);
 
         String words = "dont stand there come in";
         tvSampleThought.setText(words);
         myRef.setValue(words);
 
+        // if the user clicks this button, sign them out and return to login screen
+        btnLogout.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+
+                String uid = FirebaseAuth.getInstance().getUid();
+                // sign out the user
+                FirebaseAuth.getInstance().signOut();
+                // send back to login page
+                uid = FirebaseAuth.getInstance().getUid();
+                Intent intent = new Intent(MainActivity.this, SignupActivity.class);
+                Log.i(TAG, "entered gotoLogin2");
+                startActivity(intent);
+            }
+        });
+
     }
 
+
+
+
+
+    //FirebaseAuth.getInstance().signOut();
+
+    // TESTING APP LIFELINE
     @Override
     protected void onStart() {
         super.onStart();
@@ -56,13 +81,5 @@ public class MainActivity extends AppCompatActivity {
     protected void onResume() {
         super.onResume();
         Log.i(TAG, "entered the 'onResume' function");
-    }
-
-    public void gotoLogin(View view) {
-        Log.i(TAG, "entered gotoLogin");
-        Intent intent = new Intent(this, LoginActivity.class);
-        Log.i(TAG, "entered gotoLogin2");
-        startActivity(intent);
-
     }
 }
